Performance problem with TextField solved.
`react-bootstrap`'s onChange event seems to be fired avery time a letter is typed,
not as expected (when user stops typing), so onChange was being propagated every
time the user typed a letter. That caused a performance issue, noticeable in my
developing machine (i3 processor), so I had to address it. I solved it by setting
a timeout when onChange was fired and clearing it until user stoped typing.

The solution was inspired by [this stackoverflow answer](http://stackoverflow.com/a/14042239/667599)